new approach worked of just using a single collection
